+++ src/parsing/actions/txt/txt-relative.js
@@ -24,7 +24,7 @@ function parseRelativeTextFile(tgtInputPath, ignoreErr, relativeCallback)
 {
 	var lineCountNumber = 0;													// Current line number.
 	var retrievedDataObject = parseObjects.initializeResultObject();			// Graph object.
-	var activeNumberMode = {"flagValue": -1};									// Number read mode.
+	var activeNumberMode = {flagValue: -1};										// Number read mode.
 	var canContinue = true;														// Parse successful.
 	var flaggedErrorMessage = "";												// Error text.
 	var lineStreamObject = new lineByLine(tgtInputPath);						// Opens input file and reads by line.