+++ changelog.md
@@ -1,9 +1,27 @@
 # Changelog
 
-**./src/common/value-prep.js**
-* Rewrote the following functions for simplicity:
-	* checkInputStringType
-	* checkInputObjectType
-	* checkObjectPropertyExists
-	* checkBooleanType
-	* checkInputNumberRange
+**./src/common/sub-parse/parse-objects.js**
+* Reduced the JSON definitions in these functions:
+	* defineLineOutcomeObject
+	* definePixelRowOutcomeObject
+* Removed quotation marks for JSON definition in 'defineNodeSearchObject'
+
+---
+
+**./src/common/sub-pathfinding/block-tasks.js**
+* Removed quotation marks for JSON definition in 'initializeCounterObject'
+
+---
+
+**./src/common/sub-traverse/traverse-help-tasks.js**
+* Reduced the JSON definition in 'defineCellTraverseObject' to a single line.
+
+---
+
+**./src/output/result-folder.js**
+* Removed quotation marks in 'folderOptions' JSON.
+
+---
+
+**./src/parsing/actions/txt/txt-relative.js**
+* Removed quotation marks for JSON definition in 'parseRelativeTextFile'